backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
supabase==2.0.3
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.2
python-multipart==0.0.6
